feat: extract read pretty
