+++ packages/client/src/schema-component/antd/cascader/defaultFieldNames.ts
@@ -1,0 +1,5 @@
+export const defaultFieldNames = {
+  label: 'label',
+  value: 'value',
+  children: 'children',
+};
