+++ packages/client/src/schema-component/antd/cascader/index.md
@@ -12,3 +12,11 @@ group:
 ### Cascader usage 
 
 <code src="./demos/demo1.tsx" />
+
+### Cascader usage 
+
+<code src="./demos/demo2.tsx" />
+
+### Cascader district usage 
+
+<code src="./demos/demo3.tsx" />
